requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
python-multipart==0.0.6
Pillow==10.0.1
numpy==1.24.3
bcrypt==4.0.1
python-decouple==3.8
# Add your ML framework specific requirements below
# tensorflow==2.13.0
# or torch==2.0.1